all csv uploads from an extraction provided with the same datestamp
